requirements.txt
Openai
PyMuPdf
python-dotenv
